requirements.txt
numpy==1.19.1
matplotlib==3.3.3
pillow==7.2.0
pandas==1.1.1
Shapely==1.6.4
scikit-image==0.17.2
opencv-python==4.4.0.46
